# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema de Gestion TI - API
  version: 1.0.0
  description: >
    API REST del Sistema de Gestion de TI (tickets, inventario, licencias,
    proyectos, reportes). Todas las respuestas usan el mismo envoltorio
    { success, message, errors, data, meta } y los errores HTTP devuelven
    success=false con el detalle en errors.
  contact:
    name: Equipo de TI
servers:
  - url: http://localhost:3000
    description: Entorno local (Docker)
...
paths:
...
  /api/roles:
    get:
      tags: [roles]
      summary: Listar roles activos
      operationId: listarRoles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Roles obtenidos correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaRoles'
        '401':
          description: Token ausente, invalido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '403':
          description: Requiere el permiso roles:administrar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
    post:
      tags: [roles]
      summary: Crear un rol
      operationId: crearRol
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolCreacion'
            example:
              nombre: Soporte
              permisos: [incidencias:leer, activos:leer, activos:asignar]
      responses:
        '201':
          description: Rol creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaRol'
        '400':
          description: Nombre o permisos invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '401':
          description: Token ausente, invalido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '403':
          description: Requiere el permiso roles:administrar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '409':
          description: Ya existe un rol con ese nombre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
  /api/roles/{id}:
    get:
      tags: [roles]
      summary: Obtener un rol activo
      operationId: obtenerRol
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Rol obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaRol'
        '400':
          description: Id invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '401':
          description: Token ausente, invalido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '403':
          description: Requiere el permiso roles:administrar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '404':
          description: Rol no encontrado o inactivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
    patch:
      tags: [roles]
      summary: Actualizar nombre o permisos de un rol
      operationId: actualizarRol
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolEdicion'
      responses:
        '200':
          description: Rol actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaRol'
        '400':
          description: Id, nombre o permisos invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '401':
          description: Token ausente, invalido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '403':
          description: Requiere el permiso roles:administrar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '404':
          description: Rol no encontrado o inactivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '409':
          description: Ya existe un rol con ese nombre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
    delete:
      tags: [roles]
      summary: Desactivar un rol
      description: Borrado logico; no se permite mientras el rol tenga usuarios asignados.
      operationId: eliminarRol
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Rol desactivado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaRol'
        '401':
          description: Token ausente, invalido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '403':
          description: Requiere el permiso roles:administrar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '404':
          description: Rol no encontrado o inactivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
        '409':
          description: Hay usuarios asignados a este rol
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token devuelto por POST /api/auth/login. Enviar como
        Authorization: Bearer <token>.
  schemas:
...
    RespuestaError:
      type: object
      required: [success, message, errors]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Credenciales invalidas
        errors:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/ErrorCampo'
    ErrorCampo:
      type: object
      description: Detalle de un campo que causo el error de validacion
      properties:
        campo:
          type: string
          example: estado
        valor:
          type: [string, number, 'null']
          example: BOGUS
        mensaje:
          type: string
          example: El email es obligatorio
        permitidos:
          type: array
          items:
            type: string
          description: Presente cuando el valor debe pertenecer a un enum
...
    Rol:
      type: object
      required: [id, nombre, permisos, activo]
      properties:
        id:
          type: integer
          example: 2
        nombre:
          type: string
          example: Soporte
        permisos:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Permiso'
        activo:
          type: boolean
          example: true
    Permiso:
      type: string
      enum:
        - incidencias:leer
        - activos:leer
        - activos:crear
        - activos:editar
        - activos:estado
        - activos:asignar
        - activos:eliminar
        - roles:administrar
    RolCreacion:
      type: object
      required: [nombre, permisos]
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 191
        permisos:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Permiso'
    RolEdicion:
      type: object
      minProperties: 1
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 191
        permisos:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Permiso'
        activo:
          type: boolean
          enum: [true]
          description: Use true para reactivar un rol previamente desactivado.
    RespuestaRol:
      type: object
      required: [success, message, errors, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        errors:
          type: array
          items: {}
        data:
          $ref: '#/components/schemas/Rol'
    RespuestaRoles:
      type: object
      required: [success, message, errors, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        errors:
          type: array
          items: {}
        data:
          type: array
          items:
            $ref: '#/components/schemas/Rol'
...